XSI-1546 Fix handling of NTP configuration over upgrade
Previously we were not setting `ntp-config-method` when parsing the NTP
configuration from an old version, resulting in any NTP servers that were read
being ignored.

Set the appropriate `ntp-config-method`, and also filter out any 'default' NTP
servers from the `ntp-servers` list so we only load customer specified servers.
Note this is done using a fixed list of default domains, as this has changed in
the past and may change in future, so cannot be read from the current release
config.

Signed-off-by: Alex Brett <alex.brett@cloud.com>